wip move loading location items logic out of controller
